Intergrate buckwheat client with buckwheat service
