+++ .gitignore
@@ -19,5 +19,7 @@
 *.tar.gz
 *.rar
 
+.idea
+
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
