+++ buckwheat-scrapper-client/src/main/java/com/society/buckwheat/scrapper/client/okhttp/factories/OkHttpClientFactory.java
@@ -1,0 +1,12 @@
+package com.society.buckwheat.scrapper.client.okhttp.factories;
+
+import okhttp3.OkHttpClient;
+
+public class OkHttpClientFactory {
+    private OkHttpClientFactory() {
+    }
+
+    public static OkHttpClient initialize() {
+        return new OkHttpClient.Builder().build();
+    }
+}
