+++ buckwheat-service/src/main/java/com/society/StorePricesApplication.java
@@ -1,0 +1,13 @@
+package com.society;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
+
+@SpringBootApplication(exclude = {UserDetailsServiceAutoConfiguration.class, SecurityAutoConfiguration.class})
+public class StorePricesApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(StorePricesApplication.class, args);
+    }
+}
