+++ buckwheat-service/src/main/java/com/society/api/controller/StorePricesHealthController.java
