+++ buckwheat-service/src/main/java/com/society/config/SwaggerConfiguration.java
