+++ buckwheat-service/src/main/resources/application.yml
@@ -1,0 +1,28 @@
+spring:
+  application.name: app-store-prices
+  jackson:
+    deserialization.fail-on-unknown-properties: true
+  messages:
+    basename: messages
+    use-code-as-default-message: true
+springfox.documentation.swagger.v2.path: /swagger/documentation
+
+server:
+  port: 9999
+  servlet.contextPath: /v1
+  forward-headers-strategy: native
+
+logging.level:
+  root: INFO
+  com.society: DEBUG
+
+serp.wow:
+    baseUrl: "https://api.serpwow.com/live/search"
+    token: ${SERP_WOW_TOKEN}
+    query: гречка
+    searchType: shopping
+    gl: ua
+    hl: uk
+    location: Ukraine
+    googleDomain: google.com.ua
+    page: 1
