+++ buckwheat-service/src/test/java/com/society/StorePricesHealthCheckTest.java
@@ -1,0 +1,32 @@
+package com.society;
+
+import com.society.service.StorePricesService;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+
+@WebMvcTest
+class StorePricesHealthCheckTest {
+    @MockBean
+    StorePricesService service;
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    void shouldBeAlive() throws Exception {
+        this.mockMvc.perform(get("/"))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(content().string(containsString("Hello from Store Prices App!")));
+    }
+}
